+++ src/main/java/com/fastfoodsm/fastfood/model/Order.java
@@ -43,7 +43,7 @@ public class Order implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "date")
     @Temporal(TemporalType.DATE)
     private Date date;
@@ -72,15 +72,15 @@ public class Order implements Serializable {
 
     public void setStatus(String status) { this.status = status; }
 
-    public Order(Integer id) {
+    public Order(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
